chore(api): Remove environmentId from auth JWT
This is preparatory work for the cloud auth provider migration in order to preserve user and organization only data inside the JWT.

The current environment selection is handled separately and is persisted in local storage. As part of future work, we can store the last environment id on the user object in Novu to enable cross device memory.